requirements.txt
pygame>=2.5
websockets>=12.0
pillow>=10.0
